Conan patching: Focing some OpenSSL symbole checks to True
based on the value from a Linux build.

Since we control the OpenSSL build used, those values should always be the same.

This checks at runtime seem to work just fine on Linux and macOS, but do fail on Windows, resulting in a build break if OpenSSL >= 1.1.1 is used.

This patch should allows us to finally upgrade OpenSSL in the Windows builds from 1.0.2 to 1.1.1